+++ src/GameLogic/PlugIns/InvasionEvents/GoldenInvasionPlugIn.cs
@@ -2,7 +2,7 @@
 // Licensed under the MIT License. See LICENSE file in the project root for full license information.
 // </copyright>
 
-namespace MUnique.OpenMU.GameLogic.PlugIns;
+namespace MUnique.OpenMU.GameLogic.PlugIns.InvasionEvents;
 
 using System.Collections.Concurrent;
 using System.Diagnostics;
@@ -16,106 +16,6 @@ using MUnique.OpenMU.GameLogic.Views.World;
 using MUnique.OpenMU.PlugIns;
 
 /// <summary>
-/// Describe the state of invasion.
-/// </summary>
-public enum MobsInvasionState : byte
-{
-    /// <summary>
-    /// Not started.
-    /// </summary>
-    NotStarted,
-
-    /// <summary>
-    /// Event prepared.
-    /// </summary>
-    Prepared,
-
-    /// <summary>
-    /// Event started.
-    /// </summary>
-    Started,
-}
-
-/// <summary>
-/// Abstract configuration for periodic invasions.
-/// </summary>
-public abstract class PeriodicInvasionConfiguration
-{
-    /// <summary>
-    /// Gets or sets a value indicating whether event is enabled/disabled.
-    /// </summary>
-    public bool IsActive { get; set; } = true;
-
-    /// <summary>
-    /// Gets or sets a timetable for the event.
-    /// </summary>
-    public List<TimeOnly> Timetable { get; set; } = new(GenerateTimeSequence(TimeSpan.FromMinutes(1)));
-
-    /// <summary>
-    /// Gets or sets an event's duration.
-    /// </summary>
-    public TimeSpan EventDuration { get; set; } = TimeSpan.FromMinutes(5);
-
-    /// <summary>
-    /// Gets or sets a time delay.
-    /// </summary>
-    public TimeSpan PreStartMessageDelay { get; set; } = TimeSpan.FromSeconds(3);
-
-    /// <summary>
-    /// Gets or sets the text which prints as a golden message in the game.
-    /// </summary>
-    public string? Message { get; set; } = "Invasion's been started!";
-
-    /// <summary>
-    /// Generate a sequnce of time points like [00:00, 00:01, ...].
-    /// </summary>
-    /// <param name="duration">The duration.</param>
-    public static IEnumerable<TimeOnly> GenerateTimeSequence(TimeSpan duration)
-    {
-        var limit = TimeSpan.FromDays(1);
-        var current = TimeSpan.FromDays(0);
-
-        while (current < limit)
-        {
-            yield return TimeOnly.FromTimeSpan(current);
-            current = current.Add(duration);
-        }
-    }
-
-    /// <summary>
-    /// Check if current time is OK for starting an invasion.
-    /// </summary>
-    /// <returns>Returns true if the invasion can be started.</returns>
-    public bool IsItTimeToStartInvasion()
-    {
-        if (this.Timetable.Count == 0)
-        {
-            return true;
-        }
-
-        var nowTime = TimeOnly.FromDateTime(DateTime.UtcNow);
-        var erlier = nowTime.Add(TimeSpan.FromSeconds(-5));
-
-        // For example, p = 00:00. Check that time between 00:00:00 and 00:00:05
-        return this.Timetable.Any(p => p.IsBetween(erlier, nowTime));
-    }
-}
-
-/// <summary>
-/// Configuration for golden invasion.
-/// </summary>
-public class GoldenInvasionConfiguration : PeriodicInvasionConfiguration
-{
-    /// <summary>
-    /// Initializes a new instance of the <see cref="GoldenInvasionConfiguration"/> class.
-    /// </summary>
-    public GoldenInvasionConfiguration()
-    {
-        this.Message = "[{mapName}] Golden Invasion!";
-    }
-}
-
-/// <summary>
 /// This plugin enables Golden Invasion feature.
 /// </summary>
 [PlugIn(nameof(GoldenInvasionPlugIn), "Handle Golden Invasion event")]
@@ -124,22 +24,22 @@ public class GoldenInvasionPlugIn : IPeriodicTaskPlugIn, IObjectAddedToMapPlugIn
 {
     private class GameServerState
     {
-        public DateTime NextRunUtc { get; set; } = DateTime.UtcNow;
-
-        public MobsInvasionState State { get; set; } = MobsInvasionState.NotStarted;
-
-        public ushort MapId { get; set; } = LorenciaId;
-
-        public readonly IGameContext Context;
-
-        public GameMapDefinition Map => this.Context.Configuration.Maps.First(m => m.Number == this.MapId);
-
-        public string MapName => this.Map.Name;
-
         public GameServerState(IGameContext ctx)
         {
             this.Context = ctx;
         }
+
+        public IGameContext Context { get; private set; }
+
+        public DateTime NextRunUtc { get; set; } = DateTime.UtcNow;
+
+        public InvasionEventState State { get; set; } = InvasionEventState.NotStarted;
+
+        public ushort MapId { get; set; } = LorenciaId;
+
+        public GameMapDefinition Map => this.Context.Configuration.Maps.First(m => m.Number == this.MapId);
+
+        public string MapName => this.Map.Name;
     }
 
     private static readonly ushort LorenciaId = 0;
@@ -155,7 +55,7 @@ public class GoldenInvasionPlugIn : IPeriodicTaskPlugIn, IObjectAddedToMapPlugIn
     private static readonly ushort GoldenDragonId = 79;
     private static readonly ushort GoldenVeparId = 81;
     private static readonly ushort GoldenLizardKingId = 80;
-    private static readonly ushort GoldenWheelId = 83; // ?
+    private static readonly ushort GoldenWheelId = 83;
     private static readonly ushort GoldenTantallosId = 82;
 
     private static readonly Random _random = new();
@@ -244,7 +144,7 @@ public class GoldenInvasionPlugIn : IPeriodicTaskPlugIn, IObjectAddedToMapPlugIn
 
         switch (state.State)
         {
-            case MobsInvasionState.NotStarted:
+            case InvasionEventState.NotStarted:
                 {
                     if (!configuration.IsItTimeToStartInvasion())
                     {
@@ -253,7 +153,7 @@ public class GoldenInvasionPlugIn : IPeriodicTaskPlugIn, IObjectAddedToMapPlugIn
 
                     state.NextRunUtc = DateTime.UtcNow.Add(configuration.PreStartMessageDelay);
                     state.MapId = PossibleMaps[_random.Next(0, PossibleMaps.Length)];
-                    state.State = MobsInvasionState.Prepared;
+                    state.State = InvasionEventState.Prepared;
 
                     logger.LogInformation($"{state.MapName}: initialized");
 
@@ -270,11 +170,11 @@ public class GoldenInvasionPlugIn : IPeriodicTaskPlugIn, IObjectAddedToMapPlugIn
                     break;
                 }
 
-            case MobsInvasionState.Prepared:
+            case InvasionEventState.Prepared:
                 {
                     logger.LogInformation($"{state.MapName}: spawning mobs ...");
                     state.NextRunUtc = DateTime.UtcNow.Add(configuration.EventDuration);
-                    state.State = MobsInvasionState.Started;
+                    state.State = InvasionEventState.Started;
 
                     await this.SpawnGeneralMobsAsync(state).ConfigureAwait(false);
                     await this.SpawnUniqueMobsAsync(state).ConfigureAwait(false);
@@ -284,11 +184,11 @@ public class GoldenInvasionPlugIn : IPeriodicTaskPlugIn, IObjectAddedToMapPlugIn
                     break;
                 }
 
-            case MobsInvasionState.Started:
+            case InvasionEventState.Started:
                 {
                     logger.LogInformation($"{state.MapName}: event finished");
 
-                    state.State = MobsInvasionState.NotStarted;
+                    state.State = InvasionEventState.NotStarted;
 
                     try
                     {
@@ -314,7 +214,7 @@ public class GoldenInvasionPlugIn : IPeriodicTaskPlugIn, IObjectAddedToMapPlugIn
             {
                 var state = GetStateByGameContext(player.GameContext);
 
-                var flyingEnabled = state.State != MobsInvasionState.NotStarted;
+                var flyingEnabled = state.State != InvasionEventState.NotStarted;
 
                 await this.TrySendFlyingDragonsAsync(player, flyingEnabled).ConfigureAwait(false);
             }
